+++ cmd/crawl_bg.go
@@ -120,6 +120,7 @@ func runCrawlWorker(cmd *cobra.Command, args []string, crawlID string) error {
 
 	tag, _ := cmd.Flags().GetString("tag")
 	tw := newTagWriter(tag, pc)
+	defer tw.Close()
 
 	scraper, err := newScraper(cmd)
 	if err != nil {